pyproject.toml
[project]
name = "download-faturas-engie"
version = "0.1.0"
description = "Script para fazer o download de faturas com o API da ENGIE."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mypy>=1.19.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "types-requests>=2.32.4.20260107",
]
